requirements.txt
pycord=2.6.1
python-dotenv=1.0.1
